Stop instantiating a base object for all mocks.
Basically this aligns us with more traditional mock behaviors. Mocks
have all of their mocked class's methods, but they do not need an
instantiated version of the class. An instantiated version may be useful
when mocking or stubbing out aspects of a real object (say CPWindow
instances), and that will come soon.